- Remove "lik_dens.pdf" figure from offline_analysis visualization - Remove test for "lik_dens.pdf" from offline_analysis integration test   and its expected output - Move all code for PCA/RDA/KDE model into one module - Move all tests for PCA/RDA/KDE model into one file, separated between   internals and externals - remove previous train and inference scripts (now replace by `model.fit` and `model.predict`
TODO:
- remove or update demos
- discuss removal of "lik_dens.pdf"
