+++ bcipy/signal/model/base_model.py
