+++ bcipy/signal/model/pca_rda_kde/__init__.py
@@ -1,0 +1,5 @@
+from .pca_rda_kde import PcaRdaKdeModel
+
+__all__ = [
+    "PcaRdaKdeModel",
+]
