+++ bcipy/signal/model/pca_rda_kde/demo/demo_cv.py
@@ -1,8 +1,8 @@
 import numpy as np
-from bcipy.signal.model.mach_learning.cross_validation import cross_validation
-from bcipy.signal.model.mach_learning.classifier import RegularizedDiscriminantAnalysis
-from bcipy.signal.model.mach_learning.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
-from bcipy.signal.model.mach_learning.pipeline import Pipeline
+from bcipy.signal.model.pca_rda_kde.cross_validation import cross_validation
+from bcipy.signal.model.pca_rda_kde.classifier import RegularizedDiscriminantAnalysis
+from bcipy.signal.model.pca_rda_kde.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
+from bcipy.signal.model.pca_rda_kde.pipeline import Pipeline
 
 
 def _demo_cv():
