+++ bcipy/signal/model/pca_rda_kde/demo/demo_dim_reduction.py
@@ -1,4 +1,4 @@
-from bcipy.signal.model.mach_learning.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
+from bcipy.signal.model.pca_rda_kde.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
 import numpy as np
