+++ bcipy/signal/tests/model/test_offline_analysis.py
@@ -35,7 +35,7 @@ class TestOfflineAnalysis(unittest.TestCase):
 
         cls.parameters = load_json_parameters(cls.tmp_dir / "parameters.json", value_cast=True)
         cls.model, fig_handles = offline_analysis(str(cls.tmp_dir), cls.parameters, alert_finished=False)
-        cls.mean_erp_fig_handle, cls.lik_dens_fig_handle = fig_handles
+        cls.mean_erp_fig_handle = fig_handles[0]
 
     @classmethod
     def tearDownClass(cls):
@@ -56,10 +56,6 @@ class TestOfflineAnalysis(unittest.TestCase):
     @pytest.mark.mpl_image_compare(baseline_dir=expected_output_folder, filename="test_mean_erp.png", remove_text=True)
     def test_mean_erp(self):
         return self.mean_erp_fig_handle
-
-    @pytest.mark.mpl_image_compare(baseline_dir=expected_output_folder, filename="test_lik_dens.png", remove_text=True)
-    def test_lik_dens(self):
-        return self.lik_dens_fig_handle
 
 
 if __name__ == "__main__":
