+++ bcipy/signal/tests/model/test_pca_rda_kde.py
@@ -1,3 +1,9 @@
+"""
+When the code is in a known & working state, generate "baseline" images using:
+`pytest --mpl-generate-path=bcipy/signal/tests/model/unit_test_expected_output -k "TestPcaRdaKdeModelInternals or TestPcaRdaKdeExternals"`
+"""
+import shutil
+import tempfile
 import unittest
 from pathlib import Path
 
@@ -6,79 +12,64 @@ import numpy as np
 import pytest
 from scipy.stats import iqr, norm
 
-"""Unit tests for PCA/RDA/KDE model"""
-from bcipy.signal.model.inference import inference
-from bcipy.signal.model.mach_learning.classifier import RegularizedDiscriminantAnalysis
-from bcipy.signal.model.mach_learning.cross_validation import cross_validation
-from bcipy.signal.model.mach_learning.density_estimation import KernelDensityEstimate
-from bcipy.signal.model.mach_learning.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
-from bcipy.signal.model.mach_learning.pipeline import Pipeline
-from bcipy.signal.model.mach_learning.train_model import train_pca_rda_kde_model
-
 from bcipy.helpers.task import alphabet
+from bcipy.signal.model import ModelEvaluationReport, PcaRdaKdeModel
+from bcipy.signal.model.pca_rda_kde.classifier import RegularizedDiscriminantAnalysis
+from bcipy.signal.model.pca_rda_kde.cross_validation import cross_validation
+from bcipy.signal.model.pca_rda_kde.density_estimation import KernelDensityEstimate
+from bcipy.signal.model.pca_rda_kde.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
+from bcipy.signal.model.pca_rda_kde.pipeline import Pipeline
 
 expected_output_folder = Path(__file__).absolute().parent / "unit_test_expected_output"
 
 
-class TestPcaRdaKdeModel(unittest.TestCase):
-    """
-    Note - When the code is in a known & working state, generate "baseline" images using:
-    `pytest --mpl-generate-path=bcipy/signal/tests/model/unit_test_expected_output -k TestPcaRdaKdeModel`
-    """
-
-    def setUp(self):
+class SharedSetup(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
         np.random.seed(0)
 
         # Specify data dimensions
-        self.dim_x = 10
-        self.num_channel = 8
-        self.num_x_pos = 200
-        self.num_x_neg = 200
+        cls.dim_x = 10
+        cls.num_channel = 8
+        cls.num_x_pos = 200
+        cls.num_x_neg = 200
 
         # Generate Gaussian random data
-        self.pos_mean, self.pos_std = 0, 0.5
-        self.neg_mean, self.neg_std = 1, 0.5
-        x_pos = self.pos_mean + self.pos_std * np.random.randn(self.num_channel, self.num_x_pos, self.dim_x)
-        x_neg = self.neg_mean + self.neg_std * np.random.randn(self.num_channel, self.num_x_neg, self.dim_x)
-        y_pos = np.ones(self.num_x_pos)
-        y_neg = np.zeros(self.num_x_neg)
+        cls.pos_mean, cls.pos_std = 0, 0.5
+        cls.neg_mean, cls.neg_std = 1, 0.5
+        x_pos = cls.pos_mean + cls.pos_std * np.random.randn(cls.num_channel, cls.num_x_pos, cls.dim_x)
+        x_neg = cls.neg_mean + cls.neg_std * np.random.randn(cls.num_channel, cls.num_x_neg, cls.dim_x)
+        y_pos = np.ones(cls.num_x_pos)
+        y_neg = np.zeros(cls.num_x_neg)
 
         # Stack and permute data
         x = np.concatenate([x_pos, x_neg], 1)
         y = np.concatenate([y_pos, y_neg], 0)
-        permutation = np.random.permutation(self.num_x_pos + self.num_x_neg)
+        permutation = np.random.permutation(cls.num_x_pos + cls.num_x_neg)
         x = x[:, permutation, :]
         y = y[permutation]
 
-        self.x = x
-        self.y = y
+        cls.x = x
+        cls.y = y
 
-    @pytest.mark.mpl_image_compare(
-        baseline_dir=expected_output_folder, filename="test_inference.expected.png", remove_text=True
-    )
-    def test_inference(self):
-        model, _ = train_pca_rda_kde_model(self.x, self.y, k_folds=10)
+        cls.model = PcaRdaKdeModel(k_folds=10)
+        cls.model.fit(cls.x, cls.y)
 
-        alp = alphabet()
+        cls.tmp_dir = Path(tempfile.mkdtemp())
 
-        # Create test items that resemble the fake training data, testing inference
-        num_x_p = 1
-        num_x_n = 9
+    @classmethod
+    def tearDownClass(cls):
+        shutil.rmtree(cls.tmp_dir)
 
-        x_test_pos = self.pos_mean + self.pos_std * np.random.randn(self.num_channel, num_x_p, self.dim_x)
-        x_test_neg = self.neg_mean + self.neg_std * np.random.randn(self.num_channel, num_x_n, self.dim_x)
-        x_test = np.concatenate((x_test_neg, x_test_pos), 1)
 
-        idx_let = np.random.permutation(len(alp))
-        letters = [alp[i] for i in idx_let[0 : (num_x_p + num_x_n)]]
+class TestPcaRdaKdeModelInternals(SharedSetup):
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
 
-        lik_r = inference(x=x_test, targets=letters, model=model, alphabet=alp)
-
-        fig, ax = plt.subplots()
-        ax.plot(np.array(list(range(len(alp)))), lik_r, "ro")
-        ax.set_xticks(np.arange(len(alp)))
-        ax.set_xticklabels(alp)
-        return fig
+    @classmethod
+    def tearDownClass(cls):
+        super().tearDownClass
 
     def test_pca(self):
         var_tol = 0.95
@@ -103,9 +94,8 @@ class TestPcaRdaKdeModel(unittest.TestCase):
         - TODO - can this test be re-written to use some data from self.setUp()?
           (Not vital, but would make this file shorter)
         """
-        n = 100
-
         # generate some dummy data
+        n = 100
         x = np.concatenate((np.random.normal(0, 1, int(0.3 * n)), np.random.normal(5, 1, int(0.7 * n))))[:, np.newaxis]
 
         # append 0 label to all data as we are interested in a single class case
@@ -162,8 +152,7 @@ class TestPcaRdaKdeModel(unittest.TestCase):
     def test_cv(self):
         """
         Notes:
-         - cross validation explicitly modifies pipeline[1], so we need a PCA step.
-
+        - cross validation explicitly modifies pipeline[1], so we need a PCA step.
         - The purpose of cross_validation() is to find optimal values of lambda and gamma for the RDA model
           before fitting it - it is not clear how sensitive this test is to changes in the code
           or input data, so this may be a weak test of cross_validation().
@@ -196,6 +185,61 @@ class TestPcaRdaKdeModel(unittest.TestCase):
         # output values should be correct
         expected = np.load(expected_output_folder / "test_rda.expected.npy")
         self.assertTrue(np.allclose(z, expected))
+
+
+class TestPcaRdaKdeModelExternals(SharedSetup):
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+
+    @classmethod
+    def tearDownClass(cls):
+        super().tearDownClass
+
+    @pytest.mark.mpl_image_compare(
+        baseline_dir=expected_output_folder, filename="test_inference.expected.png", remove_text=True
+    )
+    def test_fit_predict(self):
+        """Fit and then predict"""
+        alp = alphabet()
+
+        # Create test items that resemble the fake training data
+        num_x_p = 1
+        num_x_n = 9
+
+        x_test_pos = self.pos_mean + self.pos_std * np.random.randn(self.num_channel, num_x_p, self.dim_x)
+        x_test_neg = self.neg_mean + self.neg_std * np.random.randn(self.num_channel, num_x_n, self.dim_x)
+        x_test = np.concatenate((x_test_neg, x_test_pos), 1)
+
+        idx_let = np.random.permutation(len(alp))
+        letters = [alp[i] for i in idx_let[0 : (num_x_p + num_x_n)]]
+
+        lik_r = self.model.predict(data=x_test, inquiry=letters, symbol_set=alp)
+
+        fig, ax = plt.subplots()
+        ax.plot(np.array(list(range(len(alp)))), lik_r, "ro")
+        ax.set_xticks(np.arange(len(alp)))
+        ax.set_xticklabels(alp)
+        return fig
+
+    def test_evaluate(self):
+        model_report = self.model.evaluate(self.x, self.y)
+        self.assertEqual(model_report, ModelEvaluationReport(1.0))
+
+    def test_save_load(self):
+        n_trial = 15
+        symbol_set = alphabet()
+        inquiry = symbol_set[:n_trial]
+        data = np.random.randn(self.num_channel, n_trial, self.dim_x)
+        output_before = self.model.predict(data=data, inquiry=inquiry, symbol_set=symbol_set)
+
+        checkpoint_path = self.tmp_dir / "model.pkl"
+        self.model.save(checkpoint_path)
+        other_model = PcaRdaKdeModel(k_folds=self.model.k_folds)
+        other_model.load(checkpoint_path)
+        output_after = other_model.predict(data=data, inquiry=inquiry, symbol_set=symbol_set)
+
+        self.assertTrue(np.allclose(output_before, output_after))
 
 
 if __name__ == "__main__":
